requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
click==8.1.6
cmake==3.27.0
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.12
decorator==5.1.1
effdet==0.4.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gptcache==0.1.36
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonify==0.5
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
klotty==0.1.0
langchain==0.0.238
langsmith==0.0.12
layoutparser==0.3.4
lit==16.0.6
llama-index==0.5.15
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib==3.7.2
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.74
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.1
Pillow==9.5.0
portalocker==2.7.0
posthog==3.0.1
protobuf==4.23.4
pulsar-client==3.2.0
pyarrow==12.0.1
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.12.2
pypdfium2==4.18.0
PyPika==0.48.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
reliableGPT==0.3.0
requests==2.31.0
resend==0.5.1
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
scipy==1.11.1
sentencepiece==0.1.99
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
starlette==0.27.0
streamlit==1.24.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
tiktoken==0.4.0
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.8.1
unstructured-inference==0.5.5
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
validators==0.20.0
waitress==2.1.2
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
